+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ## [unreleased]
 
 ### Added
-- `sandstone/Input` prop `synchronous` to synchronize input value with onChange handler.
+- `sandstone/Input` prop `valueKey` to set input value type.
 
 ### Fixed